+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+RC
+--
+
++ jinja2 float casts are no longer just removed, but now 
+  really cast a number to a float.
+
+
 v1.0.0 
 ------